Fixed 64-bit integer truncation issues in shader translator.
This is an incompatible API change, but one which is necessary in
order to improve correctness of the code. The API version in
ShaderLang.h is updated and, unfortunately, the define renamed to
something less ambiguous due to conflicts on some Android buildbots.
Temporary patches in Chromium and WebKit will be landed separately to
support this upgrade.

BUG=403,404,405,406,407,408,409
Review URL: https://codereview.appspot.com/7300058

git-svn-id: https://angleproject.googlecode.com/svn/trunk@1826 736b8ea6-26fd-11df-bfd4-992fa37f6226

+++ src/compiler/SymbolTable.h
@@ -169,8 +169,8 @@ public:
     void setDefined() { defined = true; }
     bool isDefined() { return defined; }
 
-    int getParamCount() const { return static_cast<int>(parameters.size()); }  
-    const TParameter& getParam(int i) const { return parameters[i]; }
+    size_t getParamCount() const { return parameters.size(); }  
+    const TParameter& getParam(size_t i) const { return parameters[i]; }
 
     virtual void dump(TInfoSink &infoSink) const;
     TFunction(const TFunction&, TStructureMap& remapper);
